change dims data type from int to size_t for cuda; still some int dims in mpi

+++ include/common/array_size_2d.hpp
@@ -1,8 +1,8 @@
 #pragma once
 
 struct ArraySize2D {
-  int num_rows;
-  int num_cols;
+  size_t num_rows;
+  size_t num_cols;
 };
 
 // Non-member operator==
