+++ test/common/test_program_args/test_program_args.cpp
@@ -25,9 +25,9 @@ TEST_F(ProgramArgsTest, DefaultInit) {
 
 TEST_F(ProgramArgsTest, TestInitForMPI) {
 
-  std::vector<int> full_matrix_dim_ = std::vector<int>({6, 6});
-  std::vector<int> tile_dim_ = std::vector<int>({3, 3});
-  std::optional<std::vector<int>> maybe_subtile_dim = std::nullopt;
+  std::vector<size_t> full_matrix_dim_ = std::vector<size_t>({6, 6});
+  std::vector<size_t> tile_dim_ = std::vector<size_t>({3, 3});
+  std::optional<std::vector<size_t>> maybe_subtile_dim = std::nullopt;
   std::optional<std::string> maybe_kernel = std::nullopt;
 
 
@@ -56,9 +56,9 @@ TEST_F(ProgramArgsTest, TestInitForMPI) {
 
 TEST_F(ProgramArgsTest, TestInitForCuda) {
 
-  std::vector<int> full_matrix_dim_ = std::vector<int>({6, 6});
-  std::vector<int> tile_dim_ = std::vector<int>({6, 6});
-  std::optional<std::vector<int>> maybe_subtile_dim = std::vector<int>({3, 3});
+  std::vector<size_t> full_matrix_dim_ = std::vector<size_t>({6, 6});
+  std::vector<size_t> tile_dim_ = std::vector<size_t>({6, 6});
+  std::optional<std::vector<size_t>> maybe_subtile_dim = std::vector<size_t>({3, 3});
   std::optional<std::string> maybe_kernel = "single_tile";
