+++ test/mpi_prefix_sum/test_mpi_prefix_sum_solver.cpp
@@ -8,10 +8,10 @@
 
 class MpiPrefixSumSolverTest : public ::testing::Test {
 protected:
-  std::vector<int> full_matrix_dim_ = std::vector<int>({4, 4});
-  // std::vector<int> grid_dim_ = std::vector<int>({2, 2});
-  std::vector<int> tile_dim_ = std::vector<int>({2, 2});
-  std::optional<std::vector<int>> maybe_subtile_dim_ = std::nullopt;
+  std::vector<size_t> full_matrix_dim_ = std::vector<size_t>({4, 4});
+  // std::vector<size_t> grid_dim_ = std::vector<size_t>({2, 2});
+  std::vector<size_t> tile_dim_ = std::vector<size_t>({2, 2});
+  std::optional<std::vector<size_t>> maybe_subtile_dim_ = std::nullopt;
   std::optional<std::string> maybe_kernel_ = std::nullopt;
 
   ProgramArgs program_args_ = ProgramArgs(
